Revert "test_with inception"
