+++ Joint_Embedding/helpfulFunctions.py
