+++ Joint_Embedding/libraries.py
