--- Joint_Embedding/ministfashon/README.md
@@ -1,13 +1,0 @@
-# README
-
-## Testing Joint Embedding approach with simple data
-
-We are testing the implementation using the FashionMNIST dataset. The FashionMNIST dataset is a popular benchmarking dataset for machine learning algorithms, consisting of grayscale images of fashion items.
-
-### Results of this approach on a simple data with a simple model
-
-I used fashonMinist dataset for testing and small model for embedding.
-
-- ![Image-01](images/one_epoch_two_classes.png)
-- ![Image-02](images/15_epoch_two_classes.png)
-- ![Image-03](images/30_epoch_10_classes.png)
